Apply ResNMTF

Apply ResNMTF to data for a range of biclusters selecting the optimal number, with optional stability analysis

apply_resnmtf(
  data,
  init_f = NULL,
  init_s = NULL,
  init_g = NULL,
  k_vec = NULL,
  phi = NULL,
  xi = NULL,
  psi = NULL,
  n_iters = NULL,
  k_min = 3,
  k_max = 8,
  distance = "euclidean",
  num_repeats = 5,
  no_clusts = FALSE,
  sample_rate = 0.9,
  n_stability = 5,
  stability = TRUE,
  stab_thres = 0.4,
  remove_unstable = TRUE,
  use_parallel = TRUE
)

Arguments

data

list of n_v matrices, data to be factorised. If only one view is supplied, can be given as a matrix.

init_f

list of matrices, initialisation for F matrices

init_s

list of matrices, initialisation for S matrices

init_g

list of matrices, initialisation for G matrices

k_vec

vector of integers, number of clusters to consider in each view, default is NULL

phi

n_v x n_v matrix, default is NULL, restriction matrices for F

xi

n_v x n_v matrix, default is NULL, restriction matrices for S

psi

n_v x n_v matrix, default is NULL, restriction matrices for G

n_iters

integer, default is NULL, number of iterations to run for, otherwise will run until convergence

k_min

positive integer, default is 3, smallest value of k to be considered initially,

k_max

positive integer, default is 6, largest value of k to be considered initially,

distance

string, default is "euclidean", distance metric to use within the bisilhouette score

num_repeats

integer, default is 5, number of repeats to use within stability analysis

no_clusts

boolean, default is FALSE, whether to return only the factorisation or not,

sample_rate

numeric, default is 0.9, proportion of data to sample for stability analysis,

n_stability

integer, default is 5, number of times to repeat stability analysis,

stability

boolean, default is TRUE, whether to perform stability analysis or not,

stab_thres

numeric, default is 0.4, threshold for stability analysis,

remove_unstable

boolean, default is TRUE, whether to remove unstable clusters or not

use_parallel

boolean, default is TRUE, wheather to use parallelisation, not applicable on Windows or linux machines

Value

list of results from ResNMTF, containing the following: - output_f: list of matrices, F matrices - output_s: list of matrices, S matrices - output_g: list of matrices, G matrices - Error: numeric, mean error - All_Error: numeric, all errors - bisil: numeric, bisilhouette score - row_clusters: list of matrices, row clusters - col_clusters: list of matrices, column clusters - lambda: list of vectors, lambda vectors - mu: list of vectors, mu vectors

Examples

data <- list(matrix(rnorm(100), nrow = 10), matrix(rnorm(100), nrow = 10))
# if it is known there are 3 clusters in each view
apply_resnmtf(data = data, k_vec = c(3, 3), n_iters = 100)
#> Warning: Matrix is not non-negative. Has been made non-negative.
#> Warning: Matrix is not non-negative. Has been made non-negative.
#> Error in loadNamespace(x): there is no package called ‘bisilhouette’
# not performing stability analysis
apply_resnmtf(data = data, stability = FALSE)
#> Warning: Matrix is not non-negative. Has been made non-negative.
#> Warning: Matrix is not non-negative. Has been made non-negative.
#> Error in loadNamespace(x): there is no package called ‘bisilhouette’
# returning only the factorisation
apply_resnmtf(data = data, no_clusts = TRUE)
#> Warning: Matrix is not non-negative. Has been made non-negative.
#> Warning: Matrix is not non-negative. Has been made non-negative.
#> Error in while (test == max_k) {    max_k <- max_k + 1    k_vec <- c(k_vec, max_k)    k <- max_k * rep(1, n_v)    new_len <- length(k_vec)    res_list[[new_len]] <- res_nmtf_inner(data, init_f, init_s,         init_g, k, phi, xi, psi, n_iters, num_repeats, distance,         no_clusts)    err_list <- c(err_list, res_list[[new_len]][["bisil"]][1])    test <- k_vec[which.max(err_list)]}: argument is of length zero
# changing the default range of k searched
apply_resnmtf(data = data, k_min = 5, k_max = 10)
#> Warning: Matrix is not non-negative. Has been made non-negative.
#> Warning: Matrix is not non-negative. Has been made non-negative.
#> Error in loadNamespace(x): there is no package called ‘bisilhouette’
# changing the distance metric
apply_resnmtf(data = data, distance = "euclidean")
#> Warning: Matrix is not non-negative. Has been made non-negative.
#> Warning: Matrix is not non-negative. Has been made non-negative.
#> Error in loadNamespace(x): there is no package called ‘bisilhouette’