Apply ResNMTF

Apply ResNMTF to data for a range of biclusters selecting the optimal number, with optional stability analysis

Usage

apply_resnmtf(
  data,
  init_f = NULL,
  init_s = NULL,
  init_g = NULL,
  k_vec = NULL,
  phi = NULL,
  xi = NULL,
  psi = NULL,
  n_iters = NULL,
  k_min = 3,
  k_max = 8,
  distance = "euclidean",
  spurious = TRUE,
  num_repeats = 5,
  no_clusts = FALSE,
  sample_rate = 0.9,
  n_stability = 5,
  stability = TRUE,
  stab_thres = 0.4,
  remove_unstable = TRUE,
  use_parallel = TRUE
)

Arguments

data

list of n_v matrices, data to be factorised. If only one view is supplied, can be given as a matrix.

init_f

list of matrices, initialisation for F matrices

init_s

list of matrices, initialisation for S matrices

init_g

list of matrices, initialisation for G matrices

k_vec

vector of integers, number of clusters to consider in each view, default is NULL

phi

n_v x n_v matrix, default is NULL, restriction matrices for F

xi

n_v x n_v matrix, default is NULL, restriction matrices for S

psi

n_v x n_v matrix, default is NULL, restriction matrices for G

n_iters

integer, default is NULL, number of iterations to run for, otherwise will run until convergence

k_min

positive integer, default is 3, smallest value of k to be considered initially,

k_max

positive integer, default is 6, largest value of k to be considered initially,

distance

string, default is "euclidean", distance metric to use within the bisilhouette score

spurious

boolean, default is TRUE, whether or not spurious biclusters should be found and removed

num_repeats

integer, default is 5, number of repeats to use within stability analysis

no_clusts

boolean, default is FALSE, whether to return only the factorisation or not,

sample_rate

numeric, default is 0.9, proportion of data to sample for stability analysis,

n_stability

integer, default is 5, number of times to repeat stability analysis,

stability

boolean, default is TRUE, whether to perform stability analysis or not,

stab_thres

numeric, default is 0.4, threshold for stability analysis,

remove_unstable

boolean, default is TRUE, whether to remove unstable clusters or not

use_parallel

boolean, default is TRUE, wheather to use parallelisation, not applicable on Windows or linux machines

Value

list of results from ResNMTF, containing the following: - output_f: list of matrices, F matrices - output_s: list of matrices, S matrices - output_g: list of matrices, G matrices - Error: numeric, mean error - All_Error: numeric, all errors - bisil: numeric, bisilhouette score - row_clusters: list of matrices, row clusters - col_clusters: list of matrices, column clusters - lambda: list of vectors, lambda vectors - mu: list of vectors, mu vectors

Examples

row_clusters <- cbind(
  rbinom(100, 1, 0.5),
  rbinom(100, 1, 0.5),
  rbinom(100, 1, 0.5)
)
col_clusters <- cbind(
  rbinom(50, 1, 0.4),
  rbinom(50, 1, 0.4),
  rbinom(50, 1, 0.4)
)
n_col <- 50
data <- list(
  row_clusters %*% diag(c(5, 5, 5)) %*% t(col_clusters) +
    abs(matrix(rnorm(100 * n_col), 100, n_col)),
  row_clusters %*% diag(c(5, 5, 5)) %*% t(col_clusters) +
    abs(0.01 * matrix(rnorm(100 * n_col), 100, n_col))
)
apply_resnmtf(data, k_max = 4)
#> $output_f
#> $output_f[[1]]
#>                 [,1]         [,2]         [,3]
#> row_1   1.211677e-03 2.102588e-02 2.799899e-03
#> row_2   1.340007e-03 1.951479e-03 8.136987e-03
#> row_3   9.389012e-04 2.373096e-03 8.054265e-03
#> row_4   1.544799e-02 1.386839e-05 1.791281e-02
#> row_5   1.722802e-02 4.530345e-04 3.328981e-03
#> row_6   1.697455e-02 1.294631e-09 7.014101e-03
#> row_7   1.566761e-02 2.109684e-02 3.135123e-03
#> row_8   1.594505e-02 2.001052e-02 1.011520e-02
#> row_9   1.786334e-02 1.950648e-02 1.263627e-03
#> row_10  1.504287e-02 1.159892e-03 1.650851e-02
#> row_11  7.245260e-16 1.060792e-03 1.438302e-02
#> row_12  1.556960e-02 1.011269e-03 1.410593e-02
#> row_13  1.700113e-02 1.639441e-16 6.697857e-03
#> row_14  1.607887e-02 7.833257e-04 1.569099e-02
#> row_15  6.468657e-04 8.317958e-04 1.716112e-02
#> row_16  1.641154e-02 2.386049e-03 1.308155e-02
#> row_17  1.574101e-02 8.935262e-06 1.548885e-02
#> row_18  1.619969e-02 2.112232e-02 8.321332e-03
#> row_19  1.755299e-02 2.725125e-12 3.254862e-03
#> row_20  1.681803e-02 1.867524e-02 3.009136e-19
#> row_21  1.682453e-03 2.137260e-02 1.811619e-03
#> row_22  1.559039e-02 4.276897e-04 1.604662e-02
#> row_23  1.553798e-02 1.952100e-02 1.265788e-02
#> row_24  9.449682e-04 2.217711e-02 1.034303e-02
#> row_25  3.956032e-06 2.187498e-03 9.093019e-03
#> row_26  1.584604e-02 2.022347e-02 1.178811e-02
#> row_27  7.971033e-04 2.251998e-02 1.484948e-02
#> row_28  1.144930e-04 2.353787e-02 1.328171e-02
#> row_29  1.311137e-07 3.656396e-03 1.556188e-02
#> row_30  1.678824e-02 1.379692e-06 5.094410e-03
#> row_31  1.589889e-02 3.337894e-04 1.414747e-02
#> row_32  1.711918e-02 2.076792e-02 6.282679e-09
#> row_33  1.685840e-02 1.753135e-14 6.551492e-03
#> row_34  1.645924e-02 3.038533e-07 1.452627e-02
#> row_35  1.599505e-02 1.989831e-02 8.942324e-03
#> row_36  1.259860e-03 2.124917e-02 5.120305e-03
#> row_37  1.486052e-05 1.370645e-03 1.963657e-02
#> row_38  1.638232e-02 1.219852e-03 1.598413e-02
#> row_39  9.774275e-04 2.132977e-02 1.525907e-02
#> row_40  1.675742e-02 6.819323e-06 8.139063e-03
#> row_41  1.537525e-02 2.327935e-02 7.661236e-03
#> row_42  8.114119e-04 2.118782e-02 2.659446e-03
#> row_43  1.568140e-02 6.386476e-04 1.404763e-02
#> row_44  2.115873e-07 2.201424e-02 1.450275e-02
#> row_45  1.702357e-02 2.151519e-02 9.942324e-03
#> row_46  8.606916e-04 2.269998e-02 1.506573e-02
#> row_47  1.599177e-02 2.106874e-02 1.166888e-02
#> row_48  1.712526e-02 5.926605e-10 4.230957e-03
#> row_49  1.530909e-02 8.157682e-06 1.550091e-02
#> row_50  2.052648e-03 2.109260e-03 6.654542e-03
#> row_51  1.596277e-02 1.839628e-05 1.702317e-02
#> row_52  6.890393e-04 2.233277e-02 1.115153e-02
#> row_53  1.803138e-03 4.552576e-04 6.133541e-03
#> row_54  4.524415e-04 2.219680e-02 1.405661e-02
#> row_55  1.658919e-02 1.091649e-11 5.751855e-03
#> row_56  1.226437e-03 2.167213e-02 2.475996e-03
#> row_57  1.099252e-03 1.343626e-03 1.817499e-02
#> row_58  1.735368e-02 1.719957e-11 4.049193e-03
#> row_59  1.729196e-02 1.985372e-02 2.132803e-03
#> row_60  1.687396e-02 4.003798e-04 1.815035e-02
#> row_61  1.576873e-02 2.004238e-02 1.159113e-02
#> row_62  1.615758e-02 6.630786e-04 1.808350e-02
#> row_63  1.575711e-02 2.033257e-02 1.152496e-02
#> row_64  1.553828e-02 7.299914e-04 1.647022e-02
#> row_65  1.670893e-02 2.107630e-02 1.285714e-03
#> row_66  1.095429e-03 2.331407e-02 2.496834e-03
#> row_67  1.214001e-03 2.206688e-02 4.340267e-03
#> row_68  1.638657e-02 3.123310e-04 1.290070e-02
#> row_69  1.498368e-02 1.342489e-04 1.869879e-02
#> row_70  8.539964e-08 4.081710e-03 1.457141e-02
#> row_71  1.070846e-03 1.086504e-03 1.009621e-02
#> row_72  1.636461e-02 1.011028e-03 1.383541e-02
#> row_73  1.362067e-05 1.396069e-03 8.737106e-03
#> row_74  1.545345e-02 2.811791e-04 1.651565e-02
#> row_75  1.622410e-02 1.572681e-03 1.551481e-02
#> row_76  1.383806e-03 2.136248e-02 2.774512e-03
#> row_77  9.892449e-06 2.164438e-02 1.487337e-02
#> row_78  1.618527e-02 8.805208e-04 1.434854e-02
#> row_79  1.718336e-02 1.927538e-02 2.323427e-03
#> row_80  5.746620e-06 2.380414e-02 1.195892e-02
#> row_81  1.734569e-02 5.945825e-08 3.876896e-03
#> row_82  8.814244e-04 1.562001e-03 6.558869e-03
#> row_83  1.668942e-02 6.239065e-07 1.638967e-02
#> row_84  9.863727e-04 2.149355e-02 4.966733e-03
#> row_85  8.087113e-04 2.293615e-03 1.836267e-02
#> row_86  4.354959e-04 9.828636e-04 1.865789e-02
#> row_87  1.593123e-02 2.021395e-02 1.240432e-02
#> row_88  1.097877e-03 1.586722e-03 1.416546e-02
#> row_89  1.651259e-02 2.100897e-02 1.073571e-02
#> row_90  1.583293e-03 5.527113e-04 1.042262e-02
#> row_91  2.140679e-03 2.041997e-02 3.910131e-03
#> row_92  1.694165e-02 1.967071e-02 9.724881e-08
#> row_93  1.682569e-02 1.920528e-02 1.254509e-02
#> row_94  1.282137e-03 1.001751e-03 8.545550e-03
#> row_95  1.800841e-02 1.978652e-02 2.579565e-04
#> row_96  1.640193e-02 2.132851e-02 5.992616e-13
#> row_97  9.011166e-05 3.053659e-03 1.399145e-02
#> row_98  1.698774e-02 1.949368e-07 4.035984e-03
#> row_99  1.106930e-05 2.359969e-02 1.090486e-02
#> row_100 1.725207e-02 1.910388e-02 2.640039e-06
#> 
#> $output_f[[2]]
#>                 [,1]         [,2]         [,3]
#> row_101 9.005369e-13 2.149774e-02 5.706994e-03
#> row_102 1.380381e-34 3.096540e-51 7.215941e-03
#> row_103 7.801551e-28 3.546999e-45 9.166970e-03
#> row_104 1.689592e-02 4.960316e-04 1.302870e-02
#> row_105 1.622144e-02 3.097120e-20 6.726167e-03
#> row_106 1.708484e-02 6.237497e-14 3.043100e-03
#> row_107 1.686821e-02 2.162351e-02 8.234515e-04
#> row_108 1.706992e-02 2.247869e-02 1.083881e-02
#> row_109 1.688188e-02 2.183983e-02 8.171088e-10
#> row_110 1.680575e-02 1.663405e-09 1.779256e-02
#> row_111 1.269793e-30 2.786406e-33 9.089750e-03
#> row_112 1.667573e-02 1.709224e-06 1.556990e-02
#> row_113 1.705305e-02 8.996425e-17 2.648394e-03
#> row_114 1.689068e-02 9.507162e-06 1.543621e-02
#> row_115 7.041636e-07 2.449359e-04 1.675811e-02
#> row_116 1.693327e-02 6.311551e-05 1.477514e-02
#> row_117 1.706988e-02 1.687764e-04 1.308168e-02
#> row_118 1.715092e-02 2.249760e-02 1.242634e-02
#> row_119 1.686619e-02 3.673111e-15 1.945270e-03
#> row_120 1.689286e-02 2.168321e-02 1.924762e-03
#> row_121 4.127606e-11 2.134287e-02 6.350578e-03
#> row_122 1.715876e-02 1.525983e-04 1.300710e-02
#> row_123 1.724872e-02 2.250119e-02 1.047308e-02
#> row_124 1.425619e-10 2.188469e-02 1.598950e-02
#> row_125 1.763795e-28 7.854196e-48 9.380600e-03
#> row_126 1.739973e-02 2.270599e-02 1.088728e-02
#> row_127 1.238876e-04 2.250318e-02 1.382102e-02
#> row_128 1.314047e-04 2.248839e-02 1.178641e-02
#> row_129 8.518157e-15 1.535879e-10 2.065148e-02
#> row_130 1.651773e-02 5.092316e-17 2.893765e-03
#> row_131 1.667155e-02 3.442172e-08 1.679235e-02
#> row_132 1.660720e-02 2.216643e-02 1.615723e-03
#> row_133 1.663236e-02 1.126171e-17 4.437399e-03
#> row_134 1.710616e-02 1.928288e-04 1.232830e-02
#> row_135 1.667062e-02 2.262769e-02 1.250084e-02
#> row_136 2.042766e-04 2.189220e-02 2.060159e-03
#> row_137 4.713275e-08 2.532975e-04 1.492757e-02
#> row_138 1.696789e-02 4.368157e-08 1.570122e-02
#> row_139 2.544621e-05 2.204028e-02 1.711884e-02
#> row_140 1.695033e-02 3.367784e-19 2.775818e-03
#> row_141 1.720174e-02 2.251804e-02 1.061099e-02
#> row_142 5.061874e-08 2.165167e-02 2.751178e-03
#> row_143 1.719462e-02 6.308409e-08 1.468657e-02
#> row_144 2.801744e-05 2.248162e-02 1.302225e-02
#> row_145 1.698827e-02 2.255303e-02 1.283299e-02
#> row_146 6.725043e-06 2.222615e-02 1.762050e-02
#> row_147 1.695978e-02 2.290824e-02 1.055466e-02
#> row_148 1.674160e-02 7.057013e-14 2.854394e-03
#> row_149 1.714398e-02 5.769979e-05 1.315546e-02
#> row_150 7.774014e-34 1.006284e-48 1.126559e-02
#> row_151 1.702313e-02 2.011253e-05 1.583660e-02
#> row_152 3.338825e-08 2.237892e-02 1.478041e-02
#> row_153 5.745626e-28 7.422968e-45 8.281405e-03
#> row_154 2.485649e-06 2.189994e-02 1.765426e-02
#> row_155 1.683384e-02 3.688449e-21 7.013335e-03
#> row_156 8.705309e-06 2.164860e-02 3.937176e-03
#> row_157 7.742255e-13 4.685528e-06 1.838253e-02
#> row_158 1.693510e-02 1.235780e-17 4.006673e-03
#> row_159 1.684096e-02 2.162817e-02 1.673623e-03
#> row_160 1.704207e-02 1.509565e-04 1.509119e-02
#> row_161 1.690279e-02 2.272938e-02 1.054261e-02
#> row_162 1.722782e-02 4.630394e-04 1.201768e-02
#> row_163 1.709440e-02 2.258976e-02 1.033532e-02
#> row_164 1.713832e-02 3.477552e-04 1.303569e-02
#> row_165 1.692435e-02 2.195880e-02 9.390737e-09
#> row_166 1.364133e-12 2.185330e-02 3.238640e-03
#> row_167 2.588668e-07 2.176525e-02 2.309485e-03
#> row_168 1.703233e-02 7.435240e-05 1.323644e-02
#> row_169 1.690127e-02 4.496698e-09 1.816476e-02
#> row_170 5.800647e-08 3.678048e-04 1.836887e-02
#> row_171 4.923418e-26 4.755353e-45 6.990862e-03
#> row_172 1.676199e-02 1.443068e-08 1.708268e-02
#> row_173 8.156349e-21 2.424072e-41 8.166020e-03
#> row_174 1.710429e-02 2.263908e-05 1.580321e-02
#> row_175 1.651146e-02 1.527348e-04 1.553146e-02
#> row_176 3.816909e-09 2.158418e-02 2.693657e-03
#> row_177 2.853988e-06 2.235025e-02 1.601286e-02
#> row_178 1.719185e-02 1.982052e-04 1.408840e-02
#> row_179 1.706725e-02 2.224671e-02 1.203667e-04
#> row_180 1.137457e-04 2.241567e-02 1.430291e-02
#> row_181 1.681414e-02 1.386640e-16 6.118138e-03
#> row_182 1.627015e-41 2.620171e-39 6.464787e-03
#> row_183 1.666665e-02 9.609901e-08 1.551698e-02
#> row_184 2.169442e-09 2.180417e-02 3.164703e-03
#> row_185 3.875288e-14 6.162018e-08 1.977136e-02
#> row_186 1.446425e-10 2.418525e-09 2.046010e-02
#> row_187 1.687998e-02 2.259123e-02 1.134965e-02
#> row_188 1.734633e-08 4.252254e-08 1.811146e-02
#> row_189 1.725462e-02 2.271174e-02 1.126435e-02
#> row_190 1.816589e-21 7.709445e-43 5.827578e-03
#> row_191 2.682358e-13 2.167459e-02 5.080738e-03
#> row_192 1.658670e-02 2.193239e-02 6.157591e-08
#> row_193 1.705560e-02 2.272226e-02 1.003102e-02
#> row_194 1.452881e-36 2.830657e-50 6.622253e-03
#> row_195 1.700193e-02 2.201376e-02 1.779457e-03
#> row_196 1.715404e-02 2.175112e-02 8.662200e-04
#> row_197 2.584422e-16 6.698668e-12 1.980976e-02
#> row_198 1.687543e-02 1.176700e-15 1.922990e-03
#> row_199 9.134338e-07 2.249018e-02 1.428740e-02
#> row_200 1.700648e-02 2.170424e-02 1.692599e-10
#> 
#> 
#> $output_s
#> $output_s[[1]]
#>              [,1]         [,2]         [,3]
#> [1,] 3.704071e+21 5.025076e+20 9.092853e+19
#> [2,] 4.978292e+19 8.918863e+21 1.528540e+21
#> [3,] 4.639926e+20 6.744858e+19 1.148260e+22
#> 
#> $output_s[[2]]
#>              [,1]         [,2]         [,3]
#> [1,] 2.328919e+21 1.119603e+19 9.168220e+18
#> [2,] 5.248313e+18 6.983013e+21 1.800450e+20
#> [3,] 3.754621e+19 9.531872e+18 4.332804e+21
#> 
#> 
#> $output_g
#> $output_g[[1]]
#>                [,1]         [,2]         [,3]
#> col_1  1.008336e-03 1.487219e-04 6.272985e-02
#> col_2  1.229215e-03 6.682469e-02 3.303412e-04
#> col_3  1.023422e-02 2.109067e-03 4.161512e-02
#> col_4  4.762382e-02 5.624757e-05 5.741051e-03
#> col_5  1.097501e-02 1.121827e-02 3.354232e-02
#> col_6  1.555761e-02 1.583521e-02 2.431239e-02
#> col_7  2.918653e-02 3.208342e-02 1.554044e-06
#> col_8  4.841036e-02 1.390891e-03 3.447026e-03
#> col_9  4.343787e-05 6.833963e-02 5.662642e-04
#> col_10 1.562144e-04 3.219395e-02 3.515890e-02
#> col_11 1.145620e-02 8.216405e-03 3.431224e-02
#> col_12 2.904747e-02 3.270764e-02 3.413397e-06
#> col_13 8.104772e-03 1.306511e-02 3.623187e-02
#> col_14 1.154495e-02 4.220943e-03 4.083245e-02
#> col_15 2.976596e-02 3.136837e-02 6.326469e-06
#> col_16 4.847571e-02 6.051415e-04 4.145265e-03
#> col_17 1.955151e-02 2.104906e-02 2.198104e-02
#> col_18 2.732982e-02 1.146895e-04 3.145133e-02
#> col_19 4.974492e-02 2.269672e-03 8.123893e-04
#> col_20 1.143325e-03 6.439146e-02 3.430378e-03
#> col_21 5.915584e-04 6.715448e-02 1.417250e-03
#> col_22 2.893803e-02 3.240192e-02 6.726038e-07
#> col_23 2.716301e-02 3.262596e-04 3.127833e-02
#> col_24 7.884405e-03 6.099759e-03 4.314868e-02
#> col_25 2.966623e-02 3.151439e-02 2.994037e-06
#> col_26 1.309108e-03 3.253045e-02 3.365444e-02
#> col_27 2.854009e-02 6.831931e-05 2.985993e-02
#> col_28 9.470586e-05 6.467898e-02 4.208816e-03
#> col_29 5.036063e-02 5.454347e-04 1.361856e-03
#> col_30 5.082417e-02 2.340347e-03 2.717942e-06
#> col_31 1.023275e-03 2.349362e-05 6.349377e-02
#> col_32 6.118587e-04 3.191640e-02 3.504191e-02
#> col_33 2.943332e-04 3.267364e-02 3.447037e-02
#> col_34 2.707810e-02 1.500493e-04 3.206182e-02
#> col_35 1.308424e-02 8.188436e-03 3.130306e-02
#> col_36 2.011054e-02 2.139602e-02 2.111886e-02
#> col_37 1.935622e-02 2.099669e-02 2.194668e-02
#> col_38 5.364073e-04 3.256574e-02 3.424706e-02
#> col_39 6.654100e-03 1.639003e-02 3.460318e-02
#> col_40 2.003045e-02 8.277361e-03 2.521253e-02
#> col_41 5.060153e-02 5.229048e-04 1.200506e-03
#> col_42 5.243642e-02 5.487367e-05 7.884961e-05
#> col_43 2.855975e-02 3.227531e-02 4.360878e-05
#> col_44 1.255440e-02 1.219875e-02 3.303249e-02
#> col_45 2.925257e-02 3.103564e-02 2.353755e-04
#> col_46 2.742857e-02 1.072557e-04 3.147339e-02
#> col_47 2.988870e-02 3.077366e-02 6.107204e-05
#> col_48 5.023368e-03 9.907379e-03 4.053700e-02
#> col_49 4.357195e-04 3.335745e-02 3.416768e-02
#> col_50 2.907814e-02 3.131998e-02 8.358564e-05
#> 
#> $output_g[[2]]
#>                 [,1]         [,2]         [,3]
#> col_51  4.016652e-03 6.102025e-06 6.346055e-02
#> col_52  8.460056e-06 7.250084e-02 6.869161e-06
#> col_53  4.807079e-04 8.663076e-03 4.741955e-02
#> col_54  5.090682e-02 1.628779e-07 2.147401e-07
#> col_55  9.124921e-03 6.667651e-03 4.040200e-02
#> col_56  8.697948e-03 6.714020e-03 3.712066e-02
#> col_57  2.873229e-02 3.134002e-02 1.848658e-05
#> col_58  5.089775e-02 1.718370e-07 2.274059e-07
#> col_59  1.062117e-05 7.244951e-02 8.600553e-06
#> col_60  2.166563e-03 3.239319e-02 3.495291e-02
#> col_61  1.468238e-02 1.212356e-02 2.680373e-02
#> col_62  2.872454e-02 3.134723e-02 1.850649e-05
#> col_63  5.325666e-03 1.433033e-02 3.705077e-02
#> col_64  1.051246e-02 1.138038e-02 3.205507e-02
#> col_65  2.872807e-02 3.134833e-02 1.816425e-05
#> col_66  5.089515e-02 1.781445e-07 2.474769e-07
#> col_67  2.021292e-02 2.038386e-02 2.191448e-02
#> col_68  2.819963e-02 3.145288e-05 3.061694e-02
#> col_69  5.089974e-02 1.835006e-07 2.291054e-07
#> col_70  9.155517e-06 7.251221e-02 6.471498e-06
#> col_71  1.090297e-05 7.246269e-02 7.735558e-06
#> col_72  2.872499e-02 3.135078e-02 1.842452e-05
#> col_73  2.818765e-02 3.237461e-05 3.062268e-02
#> col_74  9.644008e-03 6.569111e-03 3.868383e-02
#> col_75  2.873321e-02 3.133802e-02 1.886747e-05
#> col_76  2.163274e-03 3.237646e-02 3.497464e-02
#> col_77  2.818749e-02 3.161520e-05 3.062630e-02
#> col_78  9.985643e-06 7.248290e-02 6.982739e-06
#> col_79  5.089330e-02 1.903146e-07 2.415159e-07
#> col_80  5.090265e-02 1.804084e-07 2.323682e-07
#> col_81  4.012619e-03 5.508657e-06 6.347639e-02
#> col_82  2.150029e-03 3.239143e-02 3.497986e-02
#> col_83  2.154976e-03 3.238489e-02 3.497622e-02
#> col_84  2.818949e-02 3.114776e-05 3.062133e-02
#> col_85  1.209361e-02 3.234440e-03 3.644669e-02
#> col_86  2.022289e-02 2.039322e-02 2.189221e-02
#> col_87  2.021802e-02 2.038712e-02 2.190516e-02
#> col_88  2.152211e-03 3.238959e-02 3.497457e-02
#> col_89  1.152596e-02 1.149752e-02 3.366782e-02
#> col_90  1.246644e-02 6.863174e-03 3.617967e-02
#> col_91  5.090038e-02 1.805948e-07 2.210195e-07
#> col_92  5.089775e-02 2.106945e-07 2.295670e-07
#> col_93  2.872356e-02 3.134198e-02 1.921447e-05
#> col_94  7.632988e-03 3.378298e-03 4.284810e-02
#> col_95  2.873413e-02 3.133878e-02 1.791531e-05
#> col_96  2.818482e-02 3.243466e-05 3.062556e-02
#> col_97  2.872514e-02 3.135142e-02 1.851979e-05
#> col_98  9.436688e-03 8.436573e-03 3.550759e-02
#> col_99  2.159682e-03 3.236532e-02 3.498893e-02
#> col_100 2.872871e-02 3.134000e-02 1.920103e-05
#> 
#> 
#> $Error
#> [1] 0.109159
#> 
#> $All_Error
#>   [1] 0.3200179 0.2411700 0.2386886 0.2375294 0.2356901 0.2326837 0.2282663
#>   [8] 0.2229681 0.2176574 0.2128167 0.2085116 0.2046012 0.2009268 0.1973948
#>  [15] 0.1939698 0.1906360 0.1873776 0.1841829 0.1810527 0.1780016 0.1750530
#>  [22] 0.1722340 0.1695724 0.1670929 0.1648096 0.1627196 0.1607994 0.1590090
#>  [29] 0.1572989 0.1556176 0.1539176 0.1521602 0.1503202 0.1483910 0.1463874
#>  [36] 0.1443443 0.1423078 0.1403228 0.1384217 0.1366184 0.1349109 0.1332872
#>  [43] 0.1317321 0.1302324 0.1287789 0.1273679 0.1259998 0.1246791 0.1234131
#>  [50] 0.1222103 0.1210796 0.1200286 0.1190633 0.1181867 0.1173989 0.1166972
#>  [57] 0.1160765 0.1155298 0.1150492 0.1146266 0.1142540 0.1139241 0.1136305
#>  [64] 0.1133675 0.1131306 0.1129158 0.1127199 0.1125403 0.1123749 0.1122219
#>  [71] 0.1120799 0.1119476 0.1118241 0.1117085 0.1116001 0.1114982 0.1114024
#>  [78] 0.1113121 0.1112270 0.1111467 0.1110707 0.1109989 0.1109309 0.1108665
#>  [85] 0.1108054 0.1107474 0.1106923 0.1106399 0.1105900 0.1105426 0.1104974
#>  [92] 0.1104543 0.1104132 0.1103739 0.1103364 0.1103006 0.1102664 0.1102337
#>  [99] 0.1102024 0.1101724 0.1101436 0.1101161 0.1100896 0.1100642 0.1100398
#> [106] 0.1100163 0.1099936 0.1099718 0.1099507 0.1099303 0.1099105 0.1098913
#> [113] 0.1098726 0.1098545 0.1098367 0.1098194 0.1098025 0.1097859 0.1097697
#> [120] 0.1097538 0.1097383 0.1097230 0.1097081 0.1096935 0.1096793 0.1096654
#> [127] 0.1096519 0.1096387 0.1096260 0.1096137 0.1096017 0.1095902 0.1095790
#> [134] 0.1095682 0.1095578 0.1095478 0.1095381 0.1095288 0.1095197 0.1095110
#> [141] 0.1095026 0.1094944 0.1094865 0.1094789 0.1094714 0.1094642 0.1094572
#> [148] 0.1094504 0.1094437 0.1094373 0.1094310 0.1094249 0.1094189 0.1094130
#> [155] 0.1094073 0.1094018 0.1093963 0.1093910 0.1093858 0.1093807 0.1093757
#> [162] 0.1093709 0.1093661 0.1093615 0.1093569 0.1093525 0.1093481 0.1093439
#> [169] 0.1093397 0.1093357 0.1093317 0.1093279 0.1093242 0.1093205 0.1093170
#> [176] 0.1093136 0.1093103 0.1093070 0.1093039 0.1093008 0.1092979 0.1092950
#> [183] 0.1092922 0.1092894 0.1092868 0.1092841 0.1092816 0.1092791 0.1092766
#> [190] 0.1092742 0.1092719 0.1092695 0.1092672 0.1092650 0.1092627 0.1092605
#> [197] 0.1092583 0.1092561 0.1092540 0.1092518 0.1092496 0.1092474 0.1092452
#> [204] 0.1092430 0.1092407 0.1092384 0.1092360 0.1092336 0.1092312 0.1092287
#> [211] 0.1092261 0.1092235 0.1092208 0.1092181 0.1092155 0.1092129 0.1092103
#> [218] 0.1092078 0.1092054 0.1092031 0.1092010 0.1091989 0.1091970 0.1091951
#> [225] 0.1091934 0.1091916 0.1091900 0.1091884 0.1091869 0.1091854 0.1091839
#> [232] 0.1091825 0.1091811 0.1091797 0.1091783 0.1091770 0.1091757 0.1091744
#> [239] 0.1091732 0.1091719 0.1091707 0.1091695 0.1091684 0.1091672 0.1091660
#> [246] 0.1091649 0.1091638 0.1091627 0.1091616 0.1091605 0.1091595 0.1091584
#> [253] 0.1091574 0.1091564 0.1091554 0.1091544
#> 
#> $bisil
#> [1] 0.1958792
#> 
#> $row_clusters
#> $row_clusters[[1]]
#>         [,1] [,2] [,3]
#> row_1      0    1    0
#> row_2      0    0    0
#> row_3      0    0    0
#> row_4      1    0    1
#> row_5      1    0    0
#> row_6      1    0    0
#> row_7      1    1    0
#> row_8      1    1    1
#> row_9      1    1    0
#> row_10     1    0    1
#> row_11     0    0    1
#> row_12     1    0    1
#> row_13     1    0    0
#> row_14     1    0    1
#> row_15     0    0    1
#> row_16     1    0    1
#> row_17     1    0    1
#> row_18     1    1    0
#> row_19     1    0    0
#> row_20     1    1    0
#> row_21     0    1    0
#> row_22     1    0    1
#> row_23     1    1    1
#> row_24     0    1    1
#> row_25     0    0    0
#> row_26     1    1    1
#> row_27     0    1    1
#> row_28     0    1    1
#> row_29     0    0    1
#> row_30     1    0    0
#> row_31     1    0    1
#> row_32     1    1    0
#> row_33     1    0    0
#> row_34     1    0    1
#> row_35     1    1    0
#> row_36     0    1    0
#> row_37     0    0    1
#> row_38     1    0    1
#> row_39     0    1    1
#> row_40     1    0    0
#> row_41     1    1    0
#> row_42     0    1    0
#> row_43     1    0    1
#> row_44     0    1    1
#> row_45     1    1    0
#> row_46     0    1    1
#> row_47     1    1    1
#> row_48     1    0    0
#> row_49     1    0    1
#> row_50     0    0    0
#> row_51     1    0    1
#> row_52     0    1    1
#> row_53     0    0    0
#> row_54     0    1    1
#> row_55     1    0    0
#> row_56     0    1    0
#> row_57     0    0    1
#> row_58     1    0    0
#> row_59     1    1    0
#> row_60     1    0    1
#> row_61     1    1    1
#> row_62     1    0    1
#> row_63     1    1    1
#> row_64     1    0    1
#> row_65     1    1    0
#> row_66     0    1    0
#> row_67     0    1    0
#> row_68     1    0    1
#> row_69     1    0    1
#> row_70     0    0    1
#> row_71     0    0    1
#> row_72     1    0    1
#> row_73     0    0    0
#> row_74     1    0    1
#> row_75     1    0    1
#> row_76     0    1    0
#> row_77     0    1    1
#> row_78     1    0    1
#> row_79     1    1    0
#> row_80     0    1    1
#> row_81     1    0    0
#> row_82     0    0    0
#> row_83     1    0    1
#> row_84     0    1    0
#> row_85     0    0    1
#> row_86     0    0    1
#> row_87     1    1    1
#> row_88     0    0    1
#> row_89     1    1    1
#> row_90     0    0    1
#> row_91     0    1    0
#> row_92     1    1    0
#> row_93     1    1    1
#> row_94     0    0    0
#> row_95     1    1    0
#> row_96     1    1    0
#> row_97     0    0    1
#> row_98     1    0    0
#> row_99     0    1    1
#> row_100    1    1    0
#> 
#> $row_clusters[[2]]
#>         [,1] [,2] [,3]
#> row_101    0    1    0
#> row_102    0    0    0
#> row_103    0    0    0
#> row_104    1    0    1
#> row_105    1    0    0
#> row_106    1    0    0
#> row_107    1    1    0
#> row_108    1    1    1
#> row_109    1    1    0
#> row_110    1    0    1
#> row_111    0    0    0
#> row_112    1    0    1
#> row_113    1    0    0
#> row_114    1    0    1
#> row_115    0    0    1
#> row_116    1    0    1
#> row_117    1    0    1
#> row_118    1    1    1
#> row_119    1    0    0
#> row_120    1    1    0
#> row_121    0    1    0
#> row_122    1    0    1
#> row_123    1    1    1
#> row_124    0    1    1
#> row_125    0    0    0
#> row_126    1    1    1
#> row_127    0    1    1
#> row_128    0    1    1
#> row_129    0    0    1
#> row_130    1    0    0
#> row_131    1    0    1
#> row_132    1    1    0
#> row_133    1    0    0
#> row_134    1    0    1
#> row_135    1    1    1
#> row_136    0    1    0
#> row_137    0    0    1
#> row_138    1    0    1
#> row_139    0    1    1
#> row_140    1    0    0
#> row_141    1    1    1
#> row_142    0    1    0
#> row_143    1    0    1
#> row_144    0    1    1
#> row_145    1    1    1
#> row_146    0    1    1
#> row_147    1    1    1
#> row_148    1    0    0
#> row_149    1    0    1
#> row_150    0    0    1
#> row_151    1    0    1
#> row_152    0    1    1
#> row_153    0    0    0
#> row_154    0    1    1
#> row_155    1    0    0
#> row_156    0    1    0
#> row_157    0    0    1
#> row_158    1    0    0
#> row_159    1    1    0
#> row_160    1    0    1
#> row_161    1    1    1
#> row_162    1    0    1
#> row_163    1    1    1
#> row_164    1    0    1
#> row_165    1    1    0
#> row_166    0    1    0
#> row_167    0    1    0
#> row_168    1    0    1
#> row_169    1    0    1
#> row_170    0    0    1
#> row_171    0    0    0
#> row_172    1    0    1
#> row_173    0    0    0
#> row_174    1    0    1
#> row_175    1    0    1
#> row_176    0    1    0
#> row_177    0    1    1
#> row_178    1    0    1
#> row_179    1    1    0
#> row_180    0    1    1
#> row_181    1    0    0
#> row_182    0    0    0
#> row_183    1    0    1
#> row_184    0    1    0
#> row_185    0    0    1
#> row_186    0    0    1
#> row_187    1    1    1
#> row_188    0    0    1
#> row_189    1    1    1
#> row_190    0    0    0
#> row_191    0    1    0
#> row_192    1    1    0
#> row_193    1    1    1
#> row_194    0    0    0
#> row_195    1    1    0
#> row_196    1    1    0
#> row_197    0    0    1
#> row_198    1    0    0
#> row_199    0    1    1
#> row_200    1    1    0
#> 
#> 
#> $col_clusters
#> $col_clusters[[1]]
#>        [,1] [,2] [,3]
#> col_1     0    0    1
#> col_2     0    1    0
#> col_3     0    0    1
#> col_4     1    0    0
#> col_5     0    0    1
#> col_6     0    0    1
#> col_7     1    1    0
#> col_8     1    0    0
#> col_9     0    1    0
#> col_10    0    1    1
#> col_11    0    0    1
#> col_12    1    1    0
#> col_13    0    0    1
#> col_14    0    0    1
#> col_15    1    1    0
#> col_16    1    0    0
#> col_17    0    1    1
#> col_18    1    0    1
#> col_19    1    0    0
#> col_20    0    1    0
#> col_21    0    1    0
#> col_22    1    1    0
#> col_23    1    0    1
#> col_24    0    0    1
#> col_25    1    1    0
#> col_26    0    1    1
#> col_27    1    0    1
#> col_28    0    1    0
#> col_29    1    0    0
#> col_30    1    0    0
#> col_31    0    0    1
#> col_32    0    1    1
#> col_33    0    1    1
#> col_34    1    0    1
#> col_35    0    0    1
#> col_36    1    1    1
#> col_37    0    1    1
#> col_38    0    1    1
#> col_39    0    0    1
#> col_40    1    0    1
#> col_41    1    0    0
#> col_42    1    0    0
#> col_43    1    1    0
#> col_44    0    0    1
#> col_45    1    1    0
#> col_46    1    0    1
#> col_47    1    1    0
#> col_48    0    0    1
#> col_49    0    1    1
#> col_50    1    1    0
#> 
#> $col_clusters[[2]]
#>         [,1] [,2] [,3]
#> col_51     0    0    1
#> col_52     0    1    0
#> col_53     0    0    1
#> col_54     1    0    0
#> col_55     0    0    1
#> col_56     0    0    1
#> col_57     1    1    0
#> col_58     1    0    0
#> col_59     0    1    0
#> col_60     0    1    1
#> col_61     0    0    1
#> col_62     1    1    0
#> col_63     0    0    1
#> col_64     0    0    1
#> col_65     1    1    0
#> col_66     1    0    0
#> col_67     1    1    1
#> col_68     1    0    1
#> col_69     1    0    0
#> col_70     0    1    0
#> col_71     0    1    0
#> col_72     1    1    0
#> col_73     1    0    1
#> col_74     0    0    1
#> col_75     1    1    0
#> col_76     0    1    1
#> col_77     1    0    1
#> col_78     0    1    0
#> col_79     1    0    0
#> col_80     1    0    0
#> col_81     0    0    1
#> col_82     0    1    1
#> col_83     0    1    1
#> col_84     1    0    1
#> col_85     0    0    1
#> col_86     1    1    1
#> col_87     1    1    1
#> col_88     0    1    1
#> col_89     0    0    1
#> col_90     0    0    1
#> col_91     1    0    0
#> col_92     1    0    0
#> col_93     1    1    0
#> col_94     0    0    1
#> col_95     1    1    0
#> col_96     1    0    1
#> col_97     1    1    0
#> col_98     0    0    1
#> col_99     0    1    1
#> col_100    1    1    0
#> 
#> 
#> $lambda
#> $lambda[[1]]
#> [1] 0 0 0
#> 
#> $lambda[[2]]
#> [1] 0 0 0
#> 
#> 
#> $mu
#> $mu[[1]]
#> [1] 0 0 0
#> 
#> $mu[[2]]
#> [1] 0 0 0
#> 
#>