Apply ResNMTF

Apply ResNMTF to data for a range of biclusters selecting the optimal number, with optional stability analysis

Usage

apply_resnmtf(
  data,
  init_f = NULL,
  init_s = NULL,
  init_g = NULL,
  k_vec = NULL,
  phi = NULL,
  xi = NULL,
  psi = NULL,
  n_iters = NULL,
  k_min = 3,
  k_max = 8,
  distance = "euclidean",
  spurious = TRUE,
  num_repeats = 5,
  no_clusts = FALSE,
  sample_rate = 0.9,
  n_stability = 5,
  stability = TRUE,
  stab_thres = 0.4,
  remove_unstable = TRUE,
  use_parallel = TRUE
)

Arguments

data

list of n_v matrices, data to be factorised. If only one view is supplied, can be given as a matrix.

init_f

list of matrices, initialisation for F matrices

init_s

list of matrices, initialisation for S matrices

init_g

list of matrices, initialisation for G matrices

k_vec

vector of integers, number of clusters to consider in each view, default is NULL

phi

n_v x n_v matrix, default is NULL, restriction matrices for F

xi

n_v x n_v matrix, default is NULL, restriction matrices for S

psi

n_v x n_v matrix, default is NULL, restriction matrices for G

n_iters

integer, default is NULL, number of iterations to run for, otherwise will run until convergence

k_min

positive integer, default is 3, smallest value of k to be considered initially,

k_max

positive integer, default is 6, largest value of k to be considered initially,

distance

string, default is "euclidean", distance metric to use within the bisilhouette score

spurious

boolean, default is TRUE, whether or not spurious biclusters should be found and removed

num_repeats

integer, default is 5, number of repeats to use within stability analysis

no_clusts

boolean, default is FALSE, whether to return only the factorisation or not,

sample_rate

numeric, default is 0.9, proportion of data to sample for stability analysis,

n_stability

integer, default is 5, number of times to repeat stability analysis,

stability

boolean, default is TRUE, whether to perform stability analysis or not,

stab_thres

numeric, default is 0.4, threshold for stability analysis,

remove_unstable

boolean, default is TRUE, whether to remove unstable clusters or not

use_parallel

boolean, default is TRUE, wheather to use parallelisation, not applicable on Windows or linux machines

Value

list of results from ResNMTF, containing the following: - output_f: list of matrices, F matrices - output_s: list of matrices, S matrices - output_g: list of matrices, G matrices - Error: numeric, mean error - All_Error: numeric, all errors - bisil: numeric, bisilhouette score - row_clusters: list of matrices, row clusters - col_clusters: list of matrices, column clusters - lambda: list of vectors, lambda vectors - mu: list of vectors, mu vectors

Examples

row_clusters <- cbind(
  rbinom(100, 1, 0.5),
  rbinom(100, 1, 0.5),
  rbinom(100, 1, 0.5)
)
col_clusters <- cbind(
  rbinom(50, 1, 0.4),
  rbinom(50, 1, 0.4),
  rbinom(50, 1, 0.4)
)
n_col <- 50
data <- list(
  row_clusters %*% diag(c(5, 5, 5)) %*% t(col_clusters) +
    abs(matrix(rnorm(100 * n_col), 100, n_col)),
  row_clusters %*% diag(c(5, 5, 5)) %*% t(col_clusters) +
    abs(0.01 * matrix(rnorm(100 * n_col), 100, n_col))
)
apply_resnmtf(data, k_max = 4)
#> $output_f
#> $output_f[[1]]
#>                 [,1]         [,2]         [,3]
#> row_1   0.0128698464 1.368593e-19 3.598008e-05
#> row_2   0.0016852079 1.262756e-03 9.371972e-04
#> row_3   0.0063271683 2.660693e-03 4.493039e-03
#> row_4   0.0009985444 1.173157e-02 9.119820e-03
#> row_5   0.0058652468 1.812080e-05 4.881743e-02
#> row_6   0.0031027390 3.991458e-03 5.294161e-02
#> row_7   0.0080177788 3.229911e-12 5.744764e-03
#> row_8   0.0136048213 9.405769e-03 8.700491e-03
#> row_9   0.0140455750 1.917972e-15 1.293098e-02
#> row_10  0.0008433847 7.099234e-03 5.687787e-03
#> row_11  0.0064330505 6.478623e-03 1.842392e-03
#> row_12  0.0008287507 6.525739e-03 6.192365e-03
#> row_13  0.0025766219 6.153728e-04 5.092994e-02
#> row_14  0.0072354176 6.155932e-02 4.897854e-02
#> row_15  0.0005827303 7.389173e-03 1.790622e-11
#> row_16  0.0017239588 1.096822e-02 9.374007e-03
#> row_17  0.0004486793 4.797566e-03 4.020649e-03
#> row_18  0.0075123298 5.622328e-03 5.478864e-03
#> row_19  0.0006593486 6.825200e-04 1.018713e-02
#> row_20  0.0153301353 1.695288e-13 1.158600e-02
#> row_21  0.0148147794 1.285319e-22 6.276923e-04
#> row_22  0.0008776655 7.242618e-03 5.681859e-03
#> row_23  0.0078375837 5.684976e-03 5.003598e-03
#> row_24  0.0647194286 4.592494e-02 2.277446e-06
#> row_25  0.0009792778 4.598376e-04 3.678934e-04
#> row_26  0.0085705764 6.589114e-03 5.584785e-03
#> row_27  0.0077688391 5.437438e-03 7.754202e-06
#> row_28  0.0158973656 1.259018e-02 1.810949e-10
#> row_29  0.0018105927 1.148157e-02 2.438596e-16
#> row_30  0.0006804685 4.585797e-04 1.089609e-02
#> row_31  0.0012707817 1.206941e-02 1.001726e-02
#> row_32  0.0085044369 3.375515e-17 6.011216e-03
#> row_33  0.0003957871 2.927685e-04 7.068050e-03
#> row_34  0.0003652683 6.522937e-03 5.482420e-03
#> row_35  0.0667634413 4.708513e-02 4.387069e-02
#> row_36  0.0053653352 5.803689e-23 1.514329e-04
#> row_37  0.0005275979 4.908177e-03 4.896728e-09
#> row_38  0.0011237464 6.300648e-03 6.061145e-03
#> row_39  0.0639417714 4.922849e-02 3.734546e-04
#> row_40  0.0038530359 5.916273e-04 4.145167e-02
#> row_41  0.0151020757 9.834398e-03 8.274616e-03
#> row_42  0.0127991127 5.625804e-17 2.500751e-04
#> row_43  0.0008858920 6.842358e-03 5.665454e-03
#> row_44  0.0709861649 4.696062e-02 1.752833e-09
#> row_45  0.0079244040 5.403297e-03 6.104531e-03
#> row_46  0.0073725818 6.318618e-03 4.417726e-08
#> row_47  0.0087830355 5.731753e-03 5.530155e-03
#> row_48  0.0048011503 1.795957e-04 4.968452e-02
#> row_49  0.0007474100 7.517746e-03 5.866382e-03
#> row_50  0.0007109613 9.899495e-05 7.083702e-04
#> row_51  0.0010317978 1.125899e-02 1.016862e-02
#> row_52  0.0159973617 1.260446e-02 6.014692e-10
#> row_53  0.0034462300 4.620988e-03 3.935386e-03
#> row_54  0.0128203720 9.757471e-03 1.489169e-08
#> row_55  0.0044466939 5.417812e-04 4.874377e-02
#> row_56  0.0666428537 3.140464e-19 4.802756e-05
#> row_57  0.0005286617 6.781364e-03 3.724888e-04
#> row_58  0.0008509357 2.106193e-04 1.015006e-02
#> row_59  0.0163502728 1.130716e-16 1.175595e-02
#> row_60  0.0006783366 6.753198e-03 6.553289e-03
#> row_61  0.0618623060 5.099924e-02 4.211356e-02
#> row_62  0.0009243148 6.313812e-03 6.009832e-03
#> row_63  0.0637419474 4.782456e-02 4.623494e-02
#> row_64  0.0060062976 6.490207e-02 4.934888e-02
#> row_65  0.0080304892 3.522271e-14 5.983799e-03
#> row_66  0.0129265105 2.768310e-24 1.202759e-04
#> row_67  0.0055287265 2.025426e-22 4.847611e-05
#> row_68  0.0007273153 5.939238e-03 5.399707e-03
#> row_69  0.0008682117 1.093050e-02 9.656845e-03
#> row_70  0.0027225106 1.363726e-02 1.427118e-15
#> row_71  0.0013099549 7.784170e-04 8.725748e-04
#> row_72  0.0008206497 6.583326e-03 6.327608e-03
#> row_73  0.0004149082 8.132294e-04 4.179396e-04
#> row_74  0.0066952728 5.519642e-02 4.435335e-02
#> row_75  0.0010848792 6.395534e-03 6.233587e-03
#> row_76  0.0085383477 5.112962e-20 1.235653e-06
#> row_77  0.0078447918 5.597372e-03 2.899937e-10
#> row_78  0.0012114198 1.239406e-02 1.220421e-02
#> row_79  0.0121975403 3.050039e-15 9.557733e-03
#> row_80  0.0143652574 9.409267e-03 4.229132e-11
#> row_81  0.0010307261 1.782533e-04 1.160703e-02
#> row_82  0.0005408674 7.015047e-04 3.184762e-04
#> row_83  0.0004899695 7.214592e-03 6.233477e-03
#> row_84  0.0075539161 2.889840e-20 3.576246e-04
#> row_85  0.0069997805 5.780074e-02 3.934975e-07
#> row_86  0.0005189228 4.604314e-03 1.502045e-07
#> row_87  0.0052849742 3.934563e-03 3.983229e-03
#> row_88  0.0009367296 7.692031e-03 1.649077e-06
#> row_89  0.0687990897 4.553617e-02 4.604508e-02
#> row_90  0.0037984104 2.061638e-03 4.186732e-03
#> row_91  0.0117811406 5.264051e-20 8.563102e-04
#> row_92  0.0130374809 1.259737e-13 9.907504e-03
#> row_93  0.0075929947 5.800015e-03 6.170819e-03
#> row_94  0.0041232967 3.000613e-03 3.637723e-03
#> row_95  0.0078287895 5.100072e-15 5.979331e-03
#> row_96  0.0072515119 3.696265e-14 5.819812e-03
#> row_97  0.0008753268 7.200208e-03 4.436489e-14
#> row_98  0.0048956364 1.280496e-06 4.576995e-02
#> row_99  0.0085446612 6.181667e-03 1.627359e-11
#> row_100 0.0075745422 1.276558e-13 5.918080e-03
#> 
#> $output_f[[2]]
#>                 [,1]         [,2]          [,3]
#> row_101 3.464112e-04 2.407396e-06  1.006381e-78
#> row_102 6.327275e-07 8.998529e-08 1.959610e-260
#> row_103 7.610447e-05 6.838433e-05  6.858603e-30
#> row_104 1.331197e-04 2.976131e-04  4.232562e-09
#> row_105 1.282868e-03 8.309768e-19  1.447345e-01
#> row_106 7.608671e-04 1.078627e-18  1.115076e-01
#> row_107 2.656812e-04 3.557117e-05  1.181833e-21
#> row_108 3.811906e-04 1.935506e-04  8.412981e-11
#> row_109 3.268900e-04 9.898522e-05  1.015268e-05
#> row_110 7.785789e-05 1.740242e-04  3.331021e-20
#> row_111 1.265266e-04 2.718429e-05  6.742722e-24
#> row_112 4.056553e-05 2.053206e-04  1.082305e-20
#> row_113 1.351749e-03 1.159263e-18  6.962737e-02
#> row_114 2.815168e-04 6.180098e-02  1.400284e-01
#> row_115 5.803917e-26 1.607959e-04  1.483587e-74
#> row_116 1.331321e-04 2.975093e-04  4.240185e-09
#> row_117 4.047796e-05 7.160842e-05  6.142353e-31
#> row_118 1.276293e-04 1.855210e-04  5.385291e-23
#> row_119 7.999578e-05 3.637281e-05  1.329777e-04
#> row_120 5.019337e-04 5.310500e-05  1.259205e-05
#> row_121 2.311125e-04 3.797245e-06  5.604793e-74
#> row_122 7.552820e-05 1.689921e-04  7.266771e-21
#> row_123 1.957820e-04 9.945027e-05  6.493798e-23
#> row_124 7.334402e-02 1.195697e-01  5.565565e-11
#> row_125 1.987331e-07 1.081624e-07 3.902464e-292
#> row_126 2.229470e-04 1.132777e-04  2.479369e-20
#> row_127 8.980965e-05 1.462798e-04  7.532700e-62
#> row_128 4.270467e-04 3.069073e-04  1.642613e-42
#> row_129 1.085372e-25 1.518139e-04  2.861314e-59
#> row_130 4.289493e-05 4.420001e-05  2.039246e-04
#> row_131 1.401907e-04 3.133090e-04  4.019057e-08
#> row_132 2.251728e-04 2.383171e-05  9.215312e-21
#> row_133 4.326033e-05 4.477318e-05  8.997595e-17
#> row_134 6.833281e-05 1.529059e-04  3.548853e-23
#> row_135 1.143554e-01 6.424573e-02  1.449796e-01
#> row_136 6.514500e-05 1.044315e-06 2.262146e-105
#> row_137 1.164913e-25 8.608594e-05  1.495815e-87
#> row_138 6.398882e-05 1.134754e-04  6.103256e-20
#> row_139 6.021820e-02 9.811807e-02  4.496676e-11
#> row_140 1.370649e-03 1.569878e-18  7.099730e-02
#> row_141 3.811512e-04 1.937015e-04  8.632269e-11
#> row_142 1.764347e-04 2.832724e-06  8.909489e-80
#> row_143 7.554059e-05 1.690533e-04  7.430241e-21
#> row_144 9.292187e-02 5.289258e-02  7.650834e-11
#> row_145 1.411305e-04 2.051444e-04  5.662635e-21
#> row_146 1.673634e-04 1.204032e-04  8.122201e-62
#> row_147 2.165306e-04 1.099509e-04  6.622468e-21
#> row_148 7.023926e-04 1.195470e-18  1.016317e-01
#> row_149 7.791039e-05 1.739236e-04  3.188882e-20
#> row_150 2.099029e-07 7.577674e-08 2.322109e-322
#> row_151 7.519213e-05 3.805015e-04  6.150541e-08
#> row_152 4.272399e-04 3.068984e-04  1.798295e-42
#> row_153 1.223049e-04 5.811588e-05  7.869765e-29
#> row_154 1.747567e-04 2.847657e-04  1.780232e-47
#> row_155 1.550449e-03 1.367378e-18  7.764476e-02
#> row_156 1.064120e-01 5.942615e-04  1.998940e-33
#> row_157 7.653615e-26 1.607853e-04  1.831284e-74
#> row_158 8.020628e-05 3.649455e-05  1.326301e-04
#> row_159 5.017236e-04 5.293073e-05  1.292733e-05
#> row_160 4.172661e-05 2.115353e-04  4.908258e-20
#> row_161 1.072302e-01 7.614844e-02  6.006915e-02
#> row_162 6.222023e-05 1.101686e-04  1.329786e-20
#> row_163 6.640353e-02 1.068398e-01  1.063842e-01
#> row_164 3.486258e-04 9.491620e-02  7.507169e-02
#> row_165 2.186692e-04 2.318129e-05  2.224110e-21
#> row_166 1.763981e-04 2.779560e-06  6.291318e-80
#> row_167 1.215716e-04 8.811700e-07 2.064445e-105
#> row_168 5.636548e-05 9.959629e-05  6.427603e-23
#> row_169 7.140556e-05 3.615780e-04  6.288917e-09
#> row_170 1.822085e-25 3.055130e-04  4.387960e-53
#> row_171 4.571152e-07 2.480759e-07 1.028369e-248
#> row_172 4.053213e-05 2.053710e-04  1.121029e-20
#> row_173 2.404946e-07 7.702113e-08 8.911443e-314
#> row_174 2.920031e-04 6.369181e-02  1.443334e-01
#> row_175 4.050753e-05 2.053892e-04  1.080071e-20
#> row_176 1.924722e-04 1.372923e-06  8.365790e-93
#> row_177 1.378155e-04 7.842006e-05  1.153666e-61
#> row_178 7.379843e-05 4.253503e-04  1.873351e-04
#> row_179 4.677034e-04 6.264544e-05  1.804846e-10
#> row_180 2.682129e-04 1.526587e-04  2.815734e-47
#> row_181 2.684822e-05 2.736795e-05  2.610575e-04
#> row_182 2.116631e-07 9.645925e-08  0.000000e+00
#> row_183 6.408650e-05 1.133924e-04  6.066957e-20
#> row_184 9.065299e-05 1.440730e-06  2.678659e-96
#> row_185 1.499522e-23 9.838444e-02  6.664448e-05
#> row_186 9.216247e-26 5.605542e-05  3.323300e-87
#> row_187 9.175484e-05 1.334097e-04  4.416586e-30
#> row_188 1.320757e-25 1.362648e-04  8.318832e-74
#> row_189 9.329531e-02 5.241027e-02  1.183432e-01
#> row_190 5.895339e-05 1.281100e-04  1.715415e-34
#> row_191 3.290539e-04 2.256536e-06  5.988351e-80
#> row_192 3.847698e-04 4.083987e-05  3.363514e-10
#> row_193 1.411499e-04 2.049780e-04  5.357139e-21
#> row_194 1.750926e-04 4.296918e-05  4.661695e-44
#> row_195 2.186164e-04 2.322167e-05  2.194490e-21
#> row_196 1.288648e-04 3.913327e-05  1.134820e-23
#> row_197 1.143788e-25 1.323154e-04  1.183221e-74
#> row_198 1.074534e-03 6.930410e-19  1.235279e-01
#> row_199 1.022416e-04 1.665336e-04  8.031172e-59
#> row_200 2.656804e-04 3.553207e-05  1.178429e-21
#> 
#> 
#> $output_s
#> $output_s[[1]]
#>            [,1]       [,2]       [,3]
#> [1,] 28.3963684  1.4054238  0.3568428
#> [2,]  0.2447091 21.5037179  0.4455741
#> [3,]  1.2113438  0.8307257 37.7280926
#> 
#> $output_s[[2]]
#>              [,1]         [,2]         [,3]
#> [1,] 1.182866e+03 6.198183e-02 2.190898e-02
#> [2,] 3.822156e-02 1.177951e+03 2.004419e-02
#> [3,] 6.453127e-01 7.193532e-04 2.278625e+03
#> 
#> 
#> $output_g
#> $output_g[[1]]
#>                [,1]         [,2]         [,3]
#> col_1  1.406319e-03 3.352698e-02 4.901007e-03
#> col_2  4.464843e-02 3.739637e-11 4.900118e-04
#> col_3  5.490010e-09 8.313009e-03 6.179206e-03
#> col_4  7.707937e-05 5.853926e-03 6.664911e-02
#> col_5  5.653344e-03 2.615233e-03 3.214173e-03
#> col_6  3.617053e-11 2.735169e-03 6.419454e-03
#> col_7  4.082425e-02 3.683212e-05 6.503594e-02
#> col_8  1.572514e-03 6.702046e-03 3.916883e-02
#> col_9  2.759014e-02 1.985397e-08 2.072811e-09
#> col_10 4.028984e-02 3.207701e-02 3.871316e-07
#> col_11 1.343877e-03 1.746630e-03 4.358451e-03
#> col_12 2.822066e-02 3.433090e-03 5.422340e-02
#> col_13 2.381311e-03 1.151396e-03 6.324916e-03
#> col_14 4.208954e-03 2.169325e-03 3.478220e-03
#> col_15 2.835918e-02 1.955677e-04 5.748902e-02
#> col_16 1.040333e-05 5.857390e-03 6.849264e-02
#> col_17 4.788849e-02 4.734006e-02 3.465242e-02
#> col_18 1.881438e-03 5.806596e-02 5.694841e-02
#> col_19 3.949442e-03 2.382409e-03 6.769778e-02
#> col_20 4.756475e-02 1.188023e-04 1.146443e-05
#> col_21 2.951407e-02 8.366676e-06 4.009730e-13
#> col_22 4.142521e-02 4.071174e-04 6.429989e-02
#> col_23 6.769936e-03 4.809067e-02 3.812666e-02
#> col_24 1.480092e-03 4.191248e-03 3.243531e-03
#> col_25 3.992658e-02 5.556165e-04 6.705862e-02
#> col_26 5.010596e-02 4.422197e-02 4.207540e-04
#> col_27 1.528853e-03 5.762820e-02 5.661660e-02
#> col_28 3.977375e-02 1.044472e-03 8.654055e-05
#> col_29 2.506803e-03 4.458656e-04 4.058011e-02
#> col_30 1.762195e-03 8.609402e-04 6.065239e-02
#> col_31 6.150104e-03 3.121939e-02 8.228631e-09
#> col_32 4.162371e-02 5.163193e-02 5.108948e-06
#> col_33 2.750097e-02 4.957276e-02 2.907143e-05
#> col_34 1.472454e-03 3.605585e-02 6.599305e-02
#> col_35 2.068053e-03 1.495021e-12 5.893806e-03
#> col_36 2.698650e-02 5.705684e-02 5.477832e-02
#> col_37 4.506798e-02 3.019357e-02 6.204568e-02
#> col_38 4.255755e-02 5.006078e-02 2.215665e-08
#> col_39 1.170169e-03 2.597149e-03 5.220720e-03
#> col_40 1.392050e-04 3.352796e-03 9.187999e-03
#> col_41 3.995666e-03 4.110821e-03 3.936236e-02
#> col_42 2.836918e-04 8.856052e-03 5.959364e-02
#> col_43 4.173054e-02 5.857970e-04 6.504456e-02
#> col_44 7.370728e-04 4.856537e-03 4.007026e-03
#> col_45 2.434463e-02 4.719952e-03 5.767587e-02
#> col_46 1.676884e-03 3.504383e-02 6.588929e-02
#> col_47 4.428414e-02 2.532903e-05 3.893093e-02
#> col_48 1.716272e-03 5.444708e-03 4.889151e-03
#> col_49 4.208830e-02 3.005751e-02 2.144690e-10
#> col_50 4.497744e-02 1.308886e-03 3.849517e-02
#> 
#> $output_g[[2]]
#>                 [,1]         [,2]         [,3]
#> col_51  2.606320e-10 4.873922e-02 4.068309e-04
#> col_52  3.351107e-02 7.603432e-04 8.296006e-51
#> col_53  1.915244e-05 7.538734e-05 3.067848e-05
#> col_54  3.225976e-04 8.602513e-13 5.050739e-02
#> col_55  1.307897e-05 1.013293e-04 1.416709e-05
#> col_56  2.900450e-05 7.060382e-12 7.448261e-05
#> col_57  4.328566e-02 8.930098e-04 4.985285e-02
#> col_58  2.397303e-04 1.324148e-12 4.017863e-02
#> col_59  2.666073e-02 6.258799e-04 1.785990e-50
#> col_60  4.336215e-02 4.909909e-02 1.711303e-10
#> col_61  2.867998e-05 9.378512e-17 6.313616e-05
#> col_62  2.670546e-02 1.096804e-03 6.435461e-02
#> col_63  2.643731e-05 8.485601e-08 4.644856e-05
#> col_64  3.819649e-05 1.732118e-17 5.161962e-05
#> col_65  2.667577e-02 1.108636e-03 6.436586e-02
#> col_66  3.456753e-04 1.169499e-12 5.050914e-02
#> col_67  3.355729e-02 8.034741e-02 3.979442e-02
#> col_68  5.582426e-05 6.152277e-02 6.531498e-02
#> col_69  3.659661e-04 9.559084e-13 5.046209e-02
#> col_70  3.351124e-02 7.474250e-04 8.596545e-51
#> col_71  2.666639e-02 5.919678e-04 2.544304e-50
#> col_72  4.329134e-02 9.063741e-04 4.984724e-02
#> col_73  7.646676e-05 7.939388e-02 4.023614e-02
#> col_74  2.086993e-05 4.928705e-09 7.327128e-05
#> col_75  4.328557e-02 9.353823e-04 4.981798e-02
#> col_76  3.357524e-02 7.968937e-02 1.118785e-10
#> col_77  6.048353e-05 6.145156e-02 6.535099e-02
#> col_78  4.329200e-02 4.650143e-04 1.627844e-50
#> col_79  2.540355e-04 1.525494e-12 4.016155e-02
#> col_80  2.026944e-04 1.028068e-12 6.519954e-02
#> col_81  2.229373e-10 4.872917e-02 4.256102e-04
#> col_82  3.357756e-02 7.974742e-02 1.215172e-10
#> col_83  2.669679e-02 6.172531e-02 1.502812e-10
#> col_84  9.162875e-05 4.891360e-02 5.058895e-02
#> col_85  1.014149e-05 1.742669e-36 4.375909e-05
#> col_86  2.671871e-02 6.210496e-02 6.468871e-02
#> col_87  4.338415e-02 4.942438e-02 5.002966e-02
#> col_88  3.361671e-02 7.965354e-02 1.033048e-10
#> col_89  4.276412e-05 6.458583e-10 9.582081e-05
#> col_90  2.202335e-05 8.207230e-30 5.042068e-05
#> col_91  2.667981e-04 1.550276e-12 4.017073e-02
#> col_92  2.018742e-04 9.957570e-13 6.521223e-02
#> col_93  4.328861e-02 9.128724e-04 4.984949e-02
#> col_94  4.933870e-05 1.149321e-11 4.133146e-05
#> col_95  2.665553e-02 1.120315e-03 6.436055e-02
#> col_96  9.725138e-05 4.890226e-02 5.057755e-02
#> col_97  3.352050e-02 1.488531e-03 3.964836e-02
#> col_98  2.644312e-05 1.091005e-31 7.297025e-05
#> col_99  4.340338e-02 4.907867e-02 1.409580e-10
#> col_100 3.350924e-02 1.524519e-03 3.962381e-02
#> 
#> 
#> $Error
#> [1] 0.006380366
#> 
#> $All_Error
#>   [1] 0.114193297 0.111396447 0.108252360 0.106209672 0.104191759 0.101821264
#>   [7] 0.098956318 0.095050954 0.089467003 0.082529351 0.075901792 0.070835773
#>  [13] 0.067231088 0.064517607 0.062275470 0.060282434 0.058431474 0.056672107
#>  [19] 0.054981396 0.053350494 0.051777205 0.050260612 0.048797447 0.047381087
#>  [25] 0.046003228 0.044656915 0.043339039 0.042051016 0.040797573 0.039584681
#>  [31] 0.038417903 0.037301737 0.036239727 0.035234831 0.034289611 0.033406000
#>  [37] 0.032584542 0.031823248 0.031116793 0.030457022 0.029835010 0.029243415
#>  [43] 0.028677488 0.028134441 0.027612171 0.027108377 0.026620250 0.026144513
#>  [49] 0.025677553 0.025215530 0.024754488 0.024290486 0.023819803 0.023339220
#>  [55] 0.022846388 0.022340255 0.021821494 0.021292840 0.020759208 0.020227473
#>  [61] 0.019705854 0.019202950 0.018726648 0.018283145 0.017876356 0.017507797
#>  [67] 0.017176883 0.016881452 0.016618341 0.016383882 0.016174288 0.015985931
#>  [73] 0.015815512 0.015660143 0.015517359 0.015385094 0.015261623 0.015145507
#>  [79] 0.015035542 0.014930707 0.014830133 0.014733065 0.014638841 0.014546875
#>  [85] 0.014456639 0.014367656 0.014279493 0.014191752 0.014104065 0.014016095
#>  [91] 0.013927526 0.013838065 0.013747440 0.013655397 0.013561703 0.013466144
#>  [97] 0.013368529 0.013268691 0.013166490 0.013061817 0.012954597 0.012844793
#> [103] 0.012732412 0.012617502 0.012500162 0.012380538 0.012258822 0.012135258
#> [109] 0.012010128 0.011883753 0.011756484 0.011628693 0.011500761 0.011373071
#> [115] 0.011245992 0.011119873 0.010995033 0.010871752 0.010750265 0.010630760
#> [121] 0.010513378 0.010398211 0.010285304 0.010174662 0.010066253 0.009960014
#> [127] 0.009855856 0.009753671 0.009653339 0.009554731 0.009457717 0.009362168
#> [133] 0.009267959 0.009174977 0.009083121 0.008992302 0.008902450 0.008813512
#> [139] 0.008725454 0.008638262 0.008551943 0.008466523 0.008382050 0.008298590
#> [145] 0.008216226 0.008135061 0.008055207 0.007976793 0.007899952 0.007824823
#> [151] 0.007751553 0.007680287 0.007611164 0.007544309 0.007479836 0.007417845
#> [157] 0.007358415 0.007301608 0.007247463 0.007195997 0.007147211 0.007101085
#> [163] 0.007057582 0.007016651 0.006978224 0.006942225 0.006908568 0.006877160
#> [169] 0.006847901 0.006820686 0.006795402 0.006771930 0.006750151 0.006729950
#> [175] 0.006711208 0.006693806 0.006677626 0.006662554 0.006648484 0.006635318
#> [181] 0.006622967 0.006611352 0.006600406 0.006590072 0.006580305 0.006571065
#> [187] 0.006562317 0.006554033 0.006546186 0.006538753 0.006531711 0.006525037
#> [193] 0.006518710 0.006512709 0.006507013 0.006501604 0.006496462 0.006491571
#> [199] 0.006486915 0.006482477 0.006478243 0.006474200 0.006470336 0.006466638
#> [205] 0.006463096 0.006459701 0.006456444 0.006453317 0.006450311 0.006447421
#> [211] 0.006444639 0.006441959 0.006439377 0.006436885 0.006434481 0.006432158
#> [217] 0.006429913 0.006427742 0.006425640 0.006423604 0.006421632 0.006419719
#> [223] 0.006417864 0.006416063 0.006414314 0.006412614 0.006410962 0.006409354
#> [229] 0.006407789 0.006406264 0.006404777 0.006403327 0.006401911 0.006400529
#> [235] 0.006399179 0.006397858 0.006396566 0.006395302 0.006394063 0.006392850
#> [241] 0.006391661 0.006390494 0.006389350 0.006388226 0.006387123 0.006386039
#> [247] 0.006384973 0.006383925 0.006382893 0.006381874 0.006380860 0.006379846
#> [253] 0.006378830 0.006377809 0.006376806 0.006375840
#> 
#> $bisil
#> [1] 0.08965189
#> 
#> $row_clusters
#> $row_clusters[[1]]
#>         [,1] [,2] [,3]
#> row_1      1    0    0
#> row_2      0    0    0
#> row_3      0    0    0
#> row_4      0    1    0
#> row_5      0    0    1
#> row_6      0    0    1
#> row_7      0    0    0
#> row_8      1    0    0
#> row_9      1    0    1
#> row_10     0    0    0
#> row_11     0    0    0
#> row_12     0    0    0
#> row_13     0    0    1
#> row_14     0    1    1
#> row_15     0    0    0
#> row_16     0    1    0
#> row_17     0    0    0
#> row_18     0    0    0
#> row_19     0    0    1
#> row_20     1    0    1
#> row_21     1    0    0
#> row_22     0    0    0
#> row_23     0    0    0
#> row_24     1    1    0
#> row_25     0    0    0
#> row_26     0    0    0
#> row_27     0    0    0
#> row_28     1    1    0
#> row_29     0    1    0
#> row_30     0    0    1
#> row_31     0    1    1
#> row_32     0    0    0
#> row_33     0    0    0
#> row_34     0    0    0
#> row_35     1    1    1
#> row_36     0    0    0
#> row_37     0    0    0
#> row_38     0    0    0
#> row_39     1    1    0
#> row_40     0    0    1
#> row_41     1    0    0
#> row_42     1    0    0
#> row_43     0    0    0
#> row_44     1    1    0
#> row_45     0    0    0
#> row_46     0    0    0
#> row_47     0    0    0
#> row_48     0    0    1
#> row_49     0    0    0
#> row_50     0    0    0
#> row_51     0    1    1
#> row_52     1    1    0
#> row_53     0    0    0
#> row_54     1    0    0
#> row_55     0    0    1
#> row_56     1    0    0
#> row_57     0    0    0
#> row_58     0    0    1
#> row_59     1    0    1
#> row_60     0    0    0
#> row_61     1    1    1
#> row_62     0    0    0
#> row_63     1    1    1
#> row_64     0    1    1
#> row_65     0    0    0
#> row_66     1    0    0
#> row_67     0    0    0
#> row_68     0    0    0
#> row_69     0    1    0
#> row_70     0    1    0
#> row_71     0    0    0
#> row_72     0    0    0
#> row_73     0    0    0
#> row_74     0    1    1
#> row_75     0    0    0
#> row_76     0    0    0
#> row_77     0    0    0
#> row_78     0    1    1
#> row_79     1    0    0
#> row_80     1    0    0
#> row_81     0    0    1
#> row_82     0    0    0
#> row_83     0    0    0
#> row_84     0    0    0
#> row_85     0    1    0
#> row_86     0    0    0
#> row_87     0    0    0
#> row_88     0    0    0
#> row_89     1    1    1
#> row_90     0    0    0
#> row_91     1    0    0
#> row_92     1    0    0
#> row_93     0    0    0
#> row_94     0    0    0
#> row_95     0    0    0
#> row_96     0    0    0
#> row_97     0    0    0
#> row_98     0    0    1
#> row_99     0    0    0
#> row_100    0    0    0
#> 
#> $row_clusters[[2]]
#>         [,1] [,2] [,3]
#> row_101    0    0    0
#> row_102    0    0    0
#> row_103    0    0    0
#> row_104    0    0    0
#> row_105    0    0    1
#> row_106    0    0    1
#> row_107    0    0    0
#> row_108    0    0    0
#> row_109    0    0    0
#> row_110    0    0    0
#> row_111    0    0    0
#> row_112    0    0    0
#> row_113    0    0    1
#> row_114    0    1    1
#> row_115    0    0    0
#> row_116    0    0    0
#> row_117    0    0    0
#> row_118    0    0    0
#> row_119    0    0    0
#> row_120    0    0    0
#> row_121    0    0    0
#> row_122    0    0    0
#> row_123    0    0    0
#> row_124    1    1    0
#> row_125    0    0    0
#> row_126    0    0    0
#> row_127    0    0    0
#> row_128    0    0    0
#> row_129    0    0    0
#> row_130    0    0    0
#> row_131    0    0    0
#> row_132    0    0    0
#> row_133    0    0    0
#> row_134    0    0    0
#> row_135    1    1    1
#> row_136    0    0    0
#> row_137    0    0    0
#> row_138    0    0    0
#> row_139    1    1    0
#> row_140    0    0    1
#> row_141    0    0    0
#> row_142    0    0    0
#> row_143    0    0    0
#> row_144    1    1    0
#> row_145    0    0    0
#> row_146    0    0    0
#> row_147    0    0    0
#> row_148    0    0    1
#> row_149    0    0    0
#> row_150    0    0    0
#> row_151    0    0    0
#> row_152    0    0    0
#> row_153    0    0    0
#> row_154    0    0    0
#> row_155    0    0    1
#> row_156    1    0    0
#> row_157    0    0    0
#> row_158    0    0    0
#> row_159    0    0    0
#> row_160    0    0    0
#> row_161    1    1    1
#> row_162    0    0    0
#> row_163    1    1    1
#> row_164    0    1    1
#> row_165    0    0    0
#> row_166    0    0    0
#> row_167    0    0    0
#> row_168    0    0    0
#> row_169    0    0    0
#> row_170    0    0    0
#> row_171    0    0    0
#> row_172    0    0    0
#> row_173    0    0    0
#> row_174    0    1    1
#> row_175    0    0    0
#> row_176    0    0    0
#> row_177    0    0    0
#> row_178    0    0    0
#> row_179    0    0    0
#> row_180    0    0    0
#> row_181    0    0    0
#> row_182    0    0    0
#> row_183    0    0    0
#> row_184    0    0    0
#> row_185    0    1    0
#> row_186    0    0    0
#> row_187    0    0    0
#> row_188    0    0    0
#> row_189    1    1    1
#> row_190    0    0    0
#> row_191    0    0    0
#> row_192    0    0    0
#> row_193    0    0    0
#> row_194    0    0    0
#> row_195    0    0    0
#> row_196    0    0    0
#> row_197    0    0    0
#> row_198    0    0    1
#> row_199    0    0    0
#> row_200    0    0    0
#> 
#> 
#> $col_clusters
#> $col_clusters[[1]]
#>        [,1] [,2] [,3]
#> col_1     0    1    0
#> col_2     1    0    0
#> col_3     0    0    0
#> col_4     0    0    1
#> col_5     0    0    0
#> col_6     0    0    0
#> col_7     1    0    1
#> col_8     0    0    1
#> col_9     1    0    0
#> col_10    1    1    0
#> col_11    0    0    0
#> col_12    1    0    1
#> col_13    0    0    0
#> col_14    0    0    0
#> col_15    1    0    1
#> col_16    0    0    1
#> col_17    1    1    1
#> col_18    0    1    1
#> col_19    0    0    1
#> col_20    1    0    0
#> col_21    1    0    0
#> col_22    1    0    1
#> col_23    0    1    1
#> col_24    0    0    0
#> col_25    1    0    1
#> col_26    1    1    0
#> col_27    0    1    1
#> col_28    1    0    0
#> col_29    0    0    1
#> col_30    0    0    1
#> col_31    0    1    0
#> col_32    1    1    0
#> col_33    1    1    0
#> col_34    0    1    1
#> col_35    0    0    0
#> col_36    1    1    1
#> col_37    1    1    1
#> col_38    1    1    0
#> col_39    0    0    0
#> col_40    0    0    0
#> col_41    0    0    1
#> col_42    0    0    1
#> col_43    1    0    1
#> col_44    0    0    0
#> col_45    1    0    1
#> col_46    0    1    1
#> col_47    1    0    1
#> col_48    0    0    0
#> col_49    1    1    0
#> col_50    1    0    1
#> 
#> $col_clusters[[2]]
#>         [,1] [,2] [,3]
#> col_51     0    1    0
#> col_52     1    0    0
#> col_53     0    0    0
#> col_54     0    0    1
#> col_55     0    0    0
#> col_56     0    0    0
#> col_57     1    0    1
#> col_58     0    0    1
#> col_59     1    0    0
#> col_60     1    1    0
#> col_61     0    0    0
#> col_62     1    0    1
#> col_63     0    0    0
#> col_64     0    0    0
#> col_65     1    0    1
#> col_66     0    0    1
#> col_67     1    1    1
#> col_68     0    1    1
#> col_69     0    0    1
#> col_70     1    0    0
#> col_71     1    0    0
#> col_72     1    0    1
#> col_73     0    1    1
#> col_74     0    0    0
#> col_75     1    0    1
#> col_76     1    1    0
#> col_77     0    1    1
#> col_78     1    0    0
#> col_79     0    0    1
#> col_80     0    0    1
#> col_81     0    1    0
#> col_82     1    1    0
#> col_83     1    1    0
#> col_84     0    1    1
#> col_85     0    0    0
#> col_86     1    1    1
#> col_87     1    1    1
#> col_88     1    1    0
#> col_89     0    0    0
#> col_90     0    0    0
#> col_91     0    0    1
#> col_92     0    0    1
#> col_93     1    0    1
#> col_94     0    0    0
#> col_95     1    0    1
#> col_96     0    1    1
#> col_97     1    0    1
#> col_98     0    0    0
#> col_99     1    1    0
#> col_100    1    0    1
#> 
#> 
#> $lambda
#> $lambda[[1]]
#> [1]  5.824580e-87 1.080197e-114  1.169443e-96
#> 
#> $lambda[[2]]
#> [1] 9.758099e-74 1.411290e-60 6.024754e-02
#> 
#> 
#> $mu
#> $mu[[1]]
#> [1] 4.219248e-66 3.509803e-92 3.687609e-38
#> 
#> $mu[[2]]
#> [1] 2.882889e-29 1.403750e-02 5.555456e+00
#> 
#>